+++ rse/main/scrapers/biotools.py
@@ -96,8 +96,9 @@ class BioToolsScraper(ScraperBase):
 
         return self.results
 
-    def create(self, database=None, config_file=None):
-        """After a scrape (whether we obtain latest or a search query) we
+    def create(self, database=None, config_file=None, **kwargs):
+        """
+        After a scrape (whether we obtain latest or a search query) we
         run create to create software repositories based on results.
         """
         from rse.main import Encyclopedia